% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_frenquencies.R
\name{aa_frequencies}
\alias{aa_frequencies}
\title{Plot aa frenquencies in a seq function}
\usage{
aa_frequencies(protein_seq)
}
\arguments{
\item{protein_seq}{Any protein seq}
}
\value{
frenquency plot
}
\description{
Plot aa frenquencies in a seq function
}
