% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_subs.R
\name{base_subs}
\alias{base_subs}
\title{Nucleotide Base Substitution}
\usage{
base_subs(dna_seq)
}
\arguments{
\item{dna_seq}{dna sequence}
}
\value{
substitutes T with U in a given dna seq to generate rna seq
}
\description{
Nucleotide Base Substitution
}
