% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_seq.R
\name{create_seq}
\alias{create_seq}
\title{Sequence creator}
\usage{
create_seq(gene_size)
}
\arguments{
\item{gene_size}{The length of the required DNA sequence}
}
\value{
A DNA sequence of the specified length
}
\description{
Creates a DNA sequence of any specified length
}
\examples{
create_seq(50)
}
