% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_codons.R
\name{split_codons}
\alias{split_codons}
\title{Split a nucleotide sequence into codons}
\usage{
split_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{A character string representing a DNA or RNA sequence.}

\item{start}{An integer indicating the starting position (1, 2, or 3) for reading codons. Default is 1.}
}
\value{
A character vector containing 3-letter codons from the input sequence.
}
\description{
Split a nucleotide sequence into codons
}
\examples{
split_codons("ATGCGATAA")
split_codons("AUGCGAUAA", start = 2)
}
