% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{Transcription}
\usage{
transcription(dna_seq)
}
\arguments{
\item{dna_seq}{dna sequence}
}
\value{
substitutes T with U in a given dna seq to generate rna seq
}
\description{
Substitute T with U
}
